removes eslint packages + configs
